API/TEST: TIDYUP Folded Client class into utils.
