+++ tests/api/features/bootstrap/FeatureContext.php
@@ -1,7 +1,6 @@
 <?php
 
 // Library used to make requests and parse xml etc
-require_once(__DIR__ . '/../../client/Client.php');
 require_once(__DIR__ . '/../../client/Record.php');
 require_once(__DIR__ . '/../../client/utils.php');
 require_once(__DIR__ . '/../../Expectation.php');
